add exceções personalizadas,e a forma de trata-las

+++ src/main/java/com/devsalmeida/techstore/service/ProdutoService.java
@@ -2,6 +2,9 @@ package com.devsalmeida.techstore.service;
 
 import com.devsalmeida.techstore.entities.Produto;
 
+import com.devsalmeida.techstore.handler.EmptyStockException;
+import com.devsalmeida.techstore.handler.ProductNotFoundException;
+import com.devsalmeida.techstore.handler.ProductSaveException;
 import com.devsalmeida.techstore.repositories.ProdutoRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -20,7 +23,7 @@ public class ProdutoService {
     public List<Produto> listarProdutos(){
         List<Produto> result = produtoRepository.findAll();
         if (result.isEmpty()){
-            throw new RuntimeException();
+            throw new EmptyStockException();
         }else {
             return result;
         }
@@ -29,24 +32,39 @@ public class ProdutoService {
 
     @Transactional
     public Produto filtrarProdutoPorcodigo(Long codigo){
-      return produtoRepository.findByCodigo(codigo);
+        Produto produto = produtoRepository.findByCodigo(codigo);
+        if (produto == null){
+            throw new ProductNotFoundException();
+        }else {
+            return produto;
+        }
+
     }
 
     @Transactional
     public Produto cadastrarProduto(Produto produto){
         if(produtoRepository.existsByNome(produto.getNome())){
-            throw new RuntimeException("Já existe um produto com este nome!");
+            throw new ProductSaveException();
         }
         return produtoRepository.save(produto);
     }
 
     @Transactional
     public Produto alterarProduto(Produto produto){
-        return produtoRepository.save(produto);
+        if (produto == null){
+            throw new ProductNotFoundException();
+        } else {
+            return produtoRepository.save(produto);
+        }
     }
 
     @Transactional
     public void deletarProduto(Produto produto){
-        produtoRepository.delete(produto);
+        if(produto == null){
+            throw new ProductNotFoundException();
+        }else {
+            produtoRepository.delete(produto);
+        }
+
     }
 }
